requirements.txt
lxml~=4.6.2
pydantic~=1.7.3
python-dateutil~=2.8.1
Flask~=1.1.2
requests~=2.25.0
urllib3~=1.26.2

isort==5.6.4
flake8==3.8.4
pytest==6.1.2
requests-mock==1.8.0